+++ src/contracts/interfaces/ISlasher.sol
@@ -18,7 +18,7 @@ interface ISlasher {
     function pendingShareScalingFactor(address operator, IStrategy strategy) external view returns (uint256);
 
     // TODO: documentation
-    function shareScalingFactorAtEpoch(address operator, IStrategy strategy, int256 epoch) external view returns (uint256);
+    function shareScalingFactorAtEpoch(address operator, IStrategy strategy, uint32 epoch) external view returns (uint256);
 
     // struct used to store information about the current state of an operator's obligations to middlewares they are serving
     struct MiddlewareTimes {
